backend_auth_flask/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
python-dotenv
PyJWT
bcrypt
Flask-Mail
gunicorn
